Correction of overdraftDebt create (has an issue finding the overdraft of the user).

+++ src/database/migrations/2-create-account.js
@@ -9,14 +9,14 @@ module.exports = {
 				type: Sequelize.INTEGER
 			},
 			cpf: {
-				type: Sequelize.BIGINT(11)
+				type: Sequelize.BIGINT
 			},
 			agency: {
-				type: Sequelize.INTEGER(10),
+				type: Sequelize.INTEGER,
 				allowNull: false
 			},
 			number: {
-				type: Sequelize.INTEGER(10),
+				type: Sequelize.INTEGER,
 				allowNull: false,
 				unique: true,
 				primaryKey: true
